Restructure directories and files
