+++ types/types.ts
@@ -1,0 +1,6 @@
+import { User } from 'firebase/auth';
+
+export type SerializedUser = Pick<
+  User,
+  'uid' | 'email' | 'displayName' | 'photoURL'
+>;
